+++ other/buildfiles/rpgenginepreparer/src/Main.java
@@ -7,7 +7,7 @@ import java.util.Collections;
 public class Main {
 
     private static FTPTools ftp = null;
-    private final static String launcherVersion = "8", playVersion = "1.11.1", createVersion = "1.11.1";
+    private final static String launcherVersion = "8", playVersion = "1.11.2", createVersion = "1.11.2";
 
     public static void main(String[] args) {
 
@@ -18,9 +18,9 @@ public class Main {
         connectToFTP();
 
         createAndUploadDocumentation();
-        createAndUploadLauncher();
-        //createAndUploadCreateOrPlay("create", createVersion);
-        //createAndUploadCreateOrPlay("play", playVersion);
+        //createAndUploadLauncher();
+        createAndUploadCreateOrPlay("create", createVersion);
+        createAndUploadCreateOrPlay("play", playVersion);
 
         cleanUp();